added logic to check if user already exists and creates a user based
