+++ routes/apiRoutes.js
@@ -8,5 +8,6 @@ const db = require("../models");
 
 module.exports=(router)=>{
     router.route("/api/users/").get(usersController.findAllusers).post(usersController.createUser);
-    router.route("/api/user/:newUseremail").get(usersController.findOneuser)
+    //CHECK TO SEE IF USER ALREADY EXISTS
+    router.route("/api/user/").get(usersController.findOneuser);
 }
